Vulkan preparation: add shader to renderer (step 12)

+++ 3dEngine/resources/shaders/modelShadowMap.geom
@@ -6,7 +6,7 @@
 #define NUMBER_SHADOW_MAPS 0
 
 uniform mat4 projectionMatrix[NUMBER_SHADOW_MAPS];
-uniform uint layersToUpdate;
+uniform int layersToUpdate;
 
 layout(triangles) in;
 layout(triangle_strip, max_vertices = MAX_VERTICES) out;
@@ -15,7 +15,7 @@ const uint POWER_TWO_TAB[13] = uint[](1, 2, 4, 8, 16, 32, 64, 128, 256, 512, 102
 
 void main() {
     for (int layer = 0; layer < NUMBER_SHADOW_MAPS; layer++) {
-        if ((layersToUpdate & POWER_TWO_TAB[layer]) != uint(0)) {
+        if ((uint(layersToUpdate) & POWER_TWO_TAB[layer]) != uint(0)) {
             gl_Layer = layer;
 
             gl_Position = projectionMatrix[layer] * gl_in[0].gl_Position;
